Move photo processing endpoint to documents router
