+++ backend/main.py
@@ -9,7 +9,6 @@ from handlers.incidents import router as incidents_router
 from handlers.materials import router as materials_router
 from handlers.objects import router as objects_router
 from handlers.subobjects import router as subobjects_router
-from handlers.others import router as others_router
 from services.db.db import create_tables
 
 create_tables()
@@ -22,7 +21,6 @@ app.include_router(checks_router)
 app.include_router(incidents_router)
 app.include_router(documents_router)
 app.include_router(materials_router)
-app.include_router(others_router)
 
 
 @app.get("/")
